backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
APScheduler==3.11.0
python-dotenv==1.1.1
requests==2.32.4
beautifulsoup4==4.13.4
pywebpush==2.0.3
